pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian-parse"
version = "0.1.0"
description = "Lightweight Obsidian vault parser — extracts wikilinks, embeds, tags, and frontmatter with a single dependency (PyYAML). No Obsidian app required."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = [
    "obsidian",
    "obsidian-md",
    "markdown",
    "wikilink",
    "knowledge-graph",
    "vault",
    "frontmatter",
    "parser",
    "graph",
    "d3",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

authors = [
    { name = "agent-hanju", email = "agent.hanju@gmail.com" },
]
dependencies = [
    "PyYAML>=6.0",
]

[project.urls]
Repository = "https://github.com/agent-hanju/obsidian-parse"
Issues = "https://github.com/agent-hanju/obsidian-parse/issues"

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "mypy>=1.20.0",
    "types-PyYAML>=6.0",
]


[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_parse"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pyright]
typeCheckingMode = "strict"
